% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL_functions.r
\name{am_sampling}
\alias{am_sampling}
\title{Assessment sampling parameters}
\usage{
am_sampling(fleets = list("LL"), years = 1990:2010, types = c("catchage",
  "catchlen", "survey", "surveyage", "surveylen", "cpue", "tagging"),
  ycurr = 2010, age_y = 2000:2010, surv_y = seq(2000, 2010, 2),
  catchlen_yrs = NULL, catchage_yrs = 2000:2010, survey_yrs = NULL,
  surveyage_yrs = NULL, surveylen_yrs = NULL, cpue_yrs = NULL,
  tagging_yrs = 2005:2009, cpue_cvpar = 0.234, cpue_cv_dist = "lognormal",
  cpue_cv_process_error = 0, surveyindexvar = 0.3, surveyagevar = 0.3,
  surveylenvar = 0, survey_dist = "lognormal",
  survey_cv_process_error = 0, catch_at_dist = "multinomial",
  catch_at_r = 1e-11, pin_diff_N_in_assessment = 1, catchage_N = list(LL =
  100), catchlen_N = list(LL = 100), proportion_mortality = 0.5, ...)
}
\arguments{
\item{fleets}{list of the fleets (sub-fisheries)}

\item{years}{vector of assessment model years}

\item{types}{vector of sample data types}

\item{ycurr}{current year}

\item{age_y}{vector of age years??? (not sure this is used)}

\item{surv_y}{survey years???}

\item{catchlen_yrs}{years of catch at length}

\item{catchage_yrs}{years of catch at age}

\item{survey_yrs}{years of surveys}

\item{surveyage_yrs}{years of survey age data}

\item{surveylen_yrs}{years of survey length data}

\item{cpue_yrs}{years of CPUE data}

\item{tagging_yrs}{years of tagging data}

\item{cpue_cvpar}{Variance estimates used in assessment (default = 0.234)}

\item{cpue_cv_dist}{Error distribution of estimated abundance index in units
used in assessment either "lognormal" (default) or "normal"}

\item{cpue_cv_process_error}{is process error applied to CPUE (default = 0 = no,
1 = yes)}

\item{surveyindexvar}{survey index variance (default = 0.3)}

\item{surveyagevar}{survey age variance (default = 0.3)}

\item{surveylenvar}{survey length variance (default = 0.0)}

\item{survey_dist}{Error distribution of survey-at-age and survey-at-length
data (default = "lognormal")}

\item{survey_cv_process_error}{If 0, no process error is applied (= no down-weighting)}

\item{catch_at_dist}{<- "multinomial"     # Error distribution of catch-at-age and catch-at-length data}

\item{catch_at_r}{Robustification parameter used in binomial likelihood
(default = 1e-11, relevant when fitted prop is close to 1 or 0)}

\item{pin_diff_N_in_assessment}{If 1, use these N for assessment, if 0, use N
from sampling (in OM) in assessment}

\item{catchage_N}{Sample size for catch at age, by sex (default = 100)}

\item{catchlen_N}{Sample size for catch at length, by sex (default = 100)}

\item{proportion_mortality}{For abundance, survey and tag-recapture
observations (default = 0.5)}

\item{...}{additional parameters (not yet implemented)}
}
\description{
Assessment sampling parameters
}
\details{
Determine years of sampling (not season specific - assumes that all seasons
with a fishery are sampled!):

1. catchage:  Age composition samples
2. catchlen:  Length composition samples
3. surveyage: Age survey
4. catchlen:  Length survey
5. index:     Index of abundance
5. tagging:   Years with tag-releases

sample_years is set up for all years of the OM. No need to define season & region
(they are fishery-specific)

Each data type probably needs to be a list and looped over them
}
