% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_processes.R
\name{calc_VBlen}
\alias{calc_VBlen}
\title{Von Bertalanffy length}
\usage{
calc_VBlen(age1, object, params1)
}
\arguments{
\item{age1}{age}

\item{object}{object to store results}

\item{params1}{von Bert parameters}
}
\description{
Growth may vary seasonally, therefore the whole
object is passed on
}
\details{
NOTE: params1 should reference names not position
}
