% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_processes.R
\name{calc_VBweight}
\alias{calc_VBweight}
\title{Von Bertalanffy weight}
\usage{
calc_VBweight(age1, object, params1, params2)
}
\arguments{
\item{age1}{age}

\item{object}{object to store results}

\item{params1}{von Bert parameters}

\item{params2}{weight length parameters}
}
\description{
Growth may vary seasonally, therefore the whole (FLQuant-like)
object is passed on
}
\details{
NOTE: params1 and params2 should reference names not position
}
