% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2_generics.R
\name{casal}
\alias{casal}
\title{CASAL assessment parameters}
\arguments{
\item{params}{assessment parameters (could replace with datass)}

\item{datass}{assessment data}

\item{tag}{tagging data}

\item{casal_path}{path to casal.exe}

\item{mpd_dat}{mpd data}

\item{mvnsamples_dat}{mvn samples}

\item{inputprefix}{CASAL input file prefix}

\item{output_log}{output log file}

\item{linux}{run on Windows (default 0) or Linux (1)}

\item{intern}{CASAL is run internally (default = TRUE) with no output sent
to the R console or it is run with output sent to R console (FALSE, use for
debugging)}

\item{skel_csl}{skeleton csl file (or location of file)}

\item{csl}{csl file (name perhaps?)}

\item{pop_csl}{CASAL population csl file (name, location???)}

\item{TAC}{Initial catch level}

\item{use_specific_ratios}{If 1, then use TAC ratios provided by 'specific_ratios'}

\item{specific_ratios}{Specific TAC ratios (if use_specific_ratios=1)}

\item{Yr_current}{current year}

\item{ssb_level_deplet}{depletion level}

\item{ssb_level_target}{target level}

\item{ref_prob_deplet}{reference depletion probability}

\item{ref_prob_target}{reference target probability}

\item{tol}{tolerance}

\item{set_seed}{switch for random number seed}

\item{rand_seed}{random seed}
}
\description{
CASAL assessment parameters
}
