% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_arrays.R
\name{create_array}
\alias{create_array}
\title{Create quant array}
\usage{
create_array(om, sampling, type, fill = 0)
}
\arguments{
\item{om}{list object storing the ages, lengths etc}

\item{sampling}{can we remove this argument}

\item{type}{switch}

\item{fill}{value to fill array with (default=0)}
}
\description{
Quant array
}
\details{
The arrays are used to store the model objects
}
