% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL2_Create_Out.r
\name{create_casal_file_out}
\alias{create_casal_file_out}
\title{Create CASAL output.csl file}
\usage{
create_casal_file_out(params, casal_path, skel_csl, csl)
}
\arguments{
\item{params}{assessment parameters (could replace with datass)}

\item{casal_path}{path to casal.exe}

\item{skel_csl}{skeleton csl file (or location of file)}

\item{csl}{csl file (name perhaps?)}
}
\description{
Create CASAL output.csl file
}
