% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL2_Create_Pop.r
\name{create_casal_file_pop}
\alias{create_casal_file_pop}
\title{Create CASAL population.csl file}
\usage{
create_casal_file_pop(params, casal_path, skel_csl, csl)
}
\arguments{
\item{params}{assessment parameters (could replace with datass)}

\item{casal_path}{path to casal.exe}

\item{skel_csl}{skeleton csl file (or location of file)}

\item{csl}{csl file (name perhaps?)}
}
\description{
Create CASAL population.csl file
}
