% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{create_move_rule}
\alias{create_move_rule}
\title{Create a movement rule in the operating model}
\usage{
create_move_rule(move_matrix, om)
}
\arguments{
\item{move_matrix}{a matrix with a row for movements between each region and
columns with names ("Name", "Sex", "Year", "Season", "Origin", "Destination",
"Ages")}

\item{om}{operating model parameters}
}
\description{
Create a movement rule in the operating model
}
\details{
The output of this function is written to om$move_rules in the original
implementation of movement. The columns Sex, Year and Season have been added
after ID. A zero is used to represent
}
