% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{expand_move_matrix}
\alias{expand_move_matrix}
\title{Expand a movement matrix}
\usage{
expand_move_matrix(move_matrix, om)
}
\arguments{
\item{move_matrix}{a matrix with a row for movements between each region and
columns with names ("Name", "Sex", "Year", "Season", "Origin", "Destination",
"Ages")}

\item{om}{operating model parameters}
}
\description{
Expand a movement matrix.
}
\details{
This function replaces 'create_move_rule' for array based movement. While the
current implementation is slow it is only used once in every iteration.
}
