% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL1_Data.r
\name{get_casal_data}
\alias{get_casal_data}
\title{Get CASAL data}
\usage{
get_casal_data(datass, Yr_current, om, sampling, obs, tag, mod)
}
\arguments{
\item{datass}{Assessment parameters and data}

\item{Yr_current}{Current year}

\item{om}{Operating model parameters}

\item{sampling}{Sampling parameters}

\item{obs}{Observations}

\item{tag}{tagging data}

\item{mod}{Results from operating model}
}
\description{
Get CASAL data
}
