% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_OM.r
\name{get_initial_pop}
\alias{get_initial_pop}
\title{Get initial population numbers and calculate SSB0}
\usage{
get_initial_pop(para, res)
}
\arguments{
\item{para}{list of model parameters.}

\item{res}{list to store model results.}
}
\description{
Get initial population numbers and calculate SSB0
}
