% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL3_Run_CASAL.r
\name{get_ssb_probabilities}
\alias{get_ssb_probabilities}
\title{Get projections for absolute and relative SSB}
\usage{
get_ssb_probabilities(abs_data, datass_yr, Catch, ssb_level_deplet,
  ssb_level_target, ref_prob_deplet, ref_prob_target)
}
\arguments{
\item{abs_data}{Input matrix with rows = trials & cols = years,
data = absolute data to be plotted as box plots}

\item{datass_yr}{datass$year}

\item{Catch}{Catch level (used in extrapolation)}

\item{ssb_level_deplet}{Reference SSB level for depletion (e.g. 0.2)}

\item{ssb_level_target}{Reference SSB level for target (e.g. 0.5 for
toothfish)}

\item{ref_prob_deplet}{Reference probability of SSB dropping below
ssb_level_deplet}

\item{ref_prob_target}{Reference measure of SSB that needs to be at/above
ssb_level_target (median). Using the median is hardwired, ref_prob_target
is not actually used.}
}
\description{
Get projections for absolute and relative SSB.
}
