% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{init_age_comp}
\alias{init_age_comp}
\title{Initial age structure at start of first year}
\usage{
init_age_comp(mu, sigma, age, M)
}
\arguments{
\item{mu}{mean recruitment}

\item{sigma}{sd/mu}

\item{age}{vector of (min_age, max_age)}

\item{M}{vector of natural mortality by age class}
}
\description{
Initial age structure at start of first year
}
\details{
Check the definition of sigma wrt to the lognormal dist
}
