% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_processes.R
\name{initial_ages}
\alias{initial_ages}
\title{Initial age structure at start of first year}
\usage{
initial_ages(method, mu = NULL, sigma = NULL, age, M = NULL,
  age_series = NULL, R_min = 100)
}
\arguments{
\item{method}{stoch - stochastic recruitment (standard), provide -
provide and include natural mortality, provide_nat without natural mortality}

\item{mu}{mean recruitment}

\item{sigma}{sd/mu}

\item{age}{vector of (min_age, max_age)}

\item{M}{vector of natural mortality by age class}

\item{age_series}{provide a series of age classes (methods 'provide' and
'user')}

\item{R_min}{minimum recruitment (default=100)}
}
\description{
Initial age structure at start of first year
}
\details{
This function replaces 'init_age_comp' which will be removed in version 0.7.

Check the definition of sigma wrt to the lognormal dist
}
\examples{
# The "stoch" method used for stochastic simulations
initial_ages(method="stoch", mu=1.5e+07, sigma=1e-01, age=c(1,30), M=rep(0.155,30))

# Alternately the recruitment series is specified and natural mortality applied
initial_ages(method="provide", age=c(1,30), M=rep(0.155,30), age_series = rep(10000,30))

# Or recruitment is specified without natural mortality
initial_ages(method="user", age=c(1,10), age_series = rep(1000,10))
}
