% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging.R
\name{iTags_M}
\alias{iTags_M}
\title{Apply natural Mortality to iTags}
\usage{
iTags_M(dat, prob)
}
\arguments{
\item{dat}{Object with individual tags}

\item{prob}{Object with probabilities (e.g. natural moratlity) by age, sex and area}
}
\description{
Apply prob (natM) to array dat (iTags)
}
