% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging.R
\name{iTags_recapture}
\alias{iTags_recapture}
\title{Recapture of individual tags}
\usage{
iTags_recapture(dat, prob, y, ss, rr, om)
}
\arguments{
\item{dat}{Object with individual tags}

\item{prob}{Recapture probability by age and sex}

\item{y}{Year}

\item{ss}{Season}

\item{rr}{Region}

\item{om}{Object with parameters for operating model}
}
\description{
Apply prob (recapture probability) to array dat (iTags)
}
