% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging.R
\name{iTags_release}
\alias{iTags_release}
\title{Release of individual tags}
\usage{
iTags_release(dat, newtags, y, i, ss, rr, om, sampling)
}
\arguments{
\item{dat}{Object with individual tags}

\item{newtags}{Vector of released tag numbers by age (prior to tag-release mortality & tag loss)}

\item{y}{Year}

\item{i}{Sex}

\item{ss}{Season}

\item{rr}{Region}

\item{om}{Object with parameters for operating model}

\item{sampling}{Object with parameters for sampling}
}
\description{
Create to array dat (iTags)
}
