% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{move_fish_dframe}
\alias{move_fish_dframe}
\title{Movement function}
\usage{
move_fish_dframe(n, move_rule, age_names, type = "pop")
}
\arguments{
\item{n}{Population or tag numbers}

\item{move_rule}{movement dataframe specifying movement by region, sex, year,
season and age classes}

\item{age_names}{vector of age class names}

\item{type}{switch to specifiy movement of population ("pop") or tags ("tag")}
}
\description{
The original movement function converted from a matrix to a dataframe
}
\details{
Note this function should only be used when movement does not vary with Sex,
Year or Season
}
