% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{move_fish_dframe_old}
\alias{move_fish_dframe_old}
\title{Movement function}
\usage{
move_fish_dframe_old(dat, om)
}
\arguments{
\item{dat}{Population or tag numbers}

\item{om}{operating model object}
}
\description{
The original movement function converted from a matrix to a dataframe
}
\details{
#' Note this function should only be used when movement does not vary with Sex,
Year or Season

* use a better selection method than dimension of array
it should be a switch with a check
}
