% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{move_init}
\alias{move_init}
\title{Initial Movement}
\usage{
move_init(n, om, equilib_yrs = 100)
}
\arguments{
\item{n}{Initial population numbers}

\item{om}{operating model object}

\item{equilib_yrs}{number of years to move the population prior to the first
model timestep (default=100).}
}
\description{
Apply movement to create initial population numbers by age, sex and area
}
\details{
Note this function should only be used when movement does not vary with Sex,
Year or Season
}
