% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{move_iTags}
\alias{move_iTags}
\title{Movement function for individual tags}
\usage{
move_iTags(dat, om)
}
\arguments{
\item{dat}{Object with individual tags}

\item{om}{operating model object}
}
\description{
Movement of individual tags, based on move_rules
}
\details{
This implementation of individual based tag movement is designed for use when
a large number of regions are specified and array based movement becomes slow
}
