% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogive.R
\name{ogive}
\alias{ogive}
\title{Ogive function}
\usage{
ogive(type, ages, params)
}
\arguments{
\item{type}{ogive type, either "logistic", "ramp", "dbnormal",
"dbplateaunormal", "dbramp", "provide"}

\item{ages}{vector of ages}

\item{params}{ogive parameters (vary with type)}
}
\description{
Provide input data for ogive
}
\details{
Add extended details to the help file

logistic =

ramp =

dbnormal =

dbnormalplateau = Input parameters (see Bull et al. 2005 - CASAL, p.48):
x <= a1: 2^[(ages-a1)/sigma.left]^2;
a1 < x <= a1+a2: a.max;
x > a1+a2: 2^[(ages-(a1+a2))/sigma.right]^2
params[1]: a1 = lower.bound for plateau
params[2]: a2 = upper.bound for plateau
params[3]: sigma_left
params[4]: sigma_right
a.max set = 1

dbramp = # params[1]: low 0 -  x value where 1st ramp starts to increase from 0
params[2]: low 1 -  x value where 1st ramp peaks at 1
params[3]: high 1 - x value where 2nd ramp starts to decrease from 1
params[4]: high 0 - x value where 2nd ramp reaches 0
}
