% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{ogive_dbnormal}
\alias{ogive_dbnormal}
\title{Double gaussian ogive}
\usage{
ogive_dbnormal(ages, params)
}
\arguments{
\item{ages}{vector of ages}

\item{params}{logistic parameters}
}
\description{
Double gaussian ogive
}
