% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{ogive_dbplateaunormal}
\alias{ogive_dbplateaunormal}
\title{Double plateau gaussian ogive}
\usage{
ogive_dbplateaunormal(ages, params)
}
\arguments{
\item{ages}{vector of ages}

\item{params}{logistic parameters}
}
\description{
Double plateau gaussian ogive
}
