% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{ogive_dbramp}
\alias{ogive_dbramp}
\title{Double ramp ogive}
\usage{
ogive_dbramp(ages, params)
}
\arguments{
\item{ages}{vector of ages}

\item{params}{logistic parameters}
}
\description{
Double ramp ogive
}
