% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{ogive_logistic}
\alias{ogive_logistic}
\title{Logistic ogive}
\usage{
ogive_logistic(ages, params)
}
\arguments{
\item{ages}{vector of ages}

\item{params}{logistic parameters}
}
\description{
Logistic ogive
}
