% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{ogive_provide}
\alias{ogive_provide}
\title{Provide input data for ogive}
\usage{
ogive_provide(ages, params)
}
\arguments{
\item{ages}{vector of ages}

\item{params}{ogive parameters}
}
\description{
Provide input data for ogive
}
