% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{ogive_ramp}
\alias{ogive_ramp}
\title{Single ramp ogive}
\usage{
ogive_ramp(ages, params)
}
\arguments{
\item{ages}{vector of ages}

\item{params}{logistic parameters}
}
\description{
Single ramp ogive
}
