% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2_generics.R
\name{om}
\alias{om}
\title{Operating model parameters}
\arguments{
\item{para}{list of model parameters.}

\item{res}{list to store model results.}

\item{tag}{tagging data}
}
\description{
Operating model parameters
}
