% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_OM.r
\name{populate_om_objects}
\alias{populate_om_objects}
\title{Populate om object with data}
\usage{
populate_om_objects(para, res)
}
\arguments{
\item{para}{list of model parameters.}

\item{res}{list to store model results.}
}
\description{
Populate om object with data
}
