% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_outputs.R
\name{rel_err}
\alias{rel_err}
\title{Extract simulation data and calculate relative error}
\usage{
rel_err(sim_data, var_names)
}
\arguments{
\item{sim_data}{list of outputs from simulations (currently csv files)}

\item{var_names}{list of operating model and assessment model parameter names
to calculate relative error from}
}
\description{
Extract simulation data and calculate relative error
}
