% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_OM.r
\name{run_annual_om}
\alias{run_annual_om}
\title{Run Annual OM loops with/without assessment}
\usage{
run_annual_om(para, res, intern = TRUE)
}
\arguments{
\item{para}{list of model parameters.}

\item{res}{list to store model results.}

\item{intern}{CASAL is run internally (default = TRUE) with no output sent
to the R console or it is run with output sent to R console (FALSE, use for
debugging)}
}
\description{
Run Annual Operating model loops with/without assessment
}
\details{
Time sequence

1. Ageing

2. Recruitment

3. Maturation

4. Migration

5. Natural and fishing mortality

6. Sampling (Age samples, length samples)

7. Tag release and recaptures
}
