% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL3_Run_CASAL.r
\name{run_casal_LLprofile}
\alias{run_casal_LLprofile}
\title{Obtain likelihood profile for B0}
\usage{
run_casal_LLprofile(nsteps = 11, stepsize = 10000, casal_path,
  inputprefix = "", est_file, output_log, profile_dat, profile_out,
  profile_log, linux = 0)
}
\arguments{
\item{nsteps}{number of MCMC steps}

\item{stepsize}{step size}

\item{casal_path}{path to casal.exe}

\item{inputprefix}{CASAL input file prefix}

\item{est_file}{estimation file}

\item{output_log}{output log file}

\item{profile_dat}{profile data}

\item{profile_out}{profile output file}

\item{profile_log}{profile log file}

\item{linux}{run on Windows (default 0) or Linux (1)}
}
\description{
Obtain likelihood profile for B0
}
