% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL3_Run_CASAL.r
\name{run_casal_mvnsamples}
\alias{run_casal_mvnsamples}
\title{Run CASAL mvn}
\usage{
run_casal_mvnsamples(casal_path, mpd_dat, output_log, mvnsamples_dat,
  nsamples = 1000)
}
\arguments{
\item{casal_path}{path to casal.exe}

\item{mpd_dat}{mpd data}

\item{output_log}{output log file}

\item{mvnsamples_dat}{mvn samples}

\item{nsamples}{number of samples (default=1000)}
}
\description{
Draw MVN samples for parameter estimates from CASAL model fit
}
\details{
Based on code written by Steve Candy and used in WG-FSA-11/24
}
