% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL3_Run_CASAL.r
\name{run_casal_proj_once}
\alias{run_casal_proj_once}
\title{Run one set of projections using MVN samples}
\usage{
run_casal_proj_once(casal_path, proj_dat, mvnsamples_dat, inputprefix = "",
  set_seed = 0, rand_seed = 0, linux = 0)
}
\arguments{
\item{casal_path}{path to casal.exe}

\item{proj_dat}{proj_dat???}

\item{mvnsamples_dat}{mvn samples}

\item{inputprefix}{CASAL input file prefix}

\item{set_seed}{set random number seed}

\item{rand_seed}{random number seed}

\item{linux}{run on Windows (default 0) or Linux (1)}
}
\description{
Run one set of projections using MVN samples (using future_catches)
}
