% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL3_Run_CASAL.r
\name{run_casal_projections}
\alias{run_casal_projections}
\title{Run CASAL projections}
\usage{
run_casal_projections(datass, TAC, Yr_current, use_specific_ratios = 0,
  specific_ratios = 1, casal_path, inputprefix, pop_csl, proj_dat,
  mvnsamples_dat, linux, ssb_level_deplet = 0.2, ssb_level_target = 0.5,
  ref_prob_deplet = 0.1, ref_prob_target = "median", tol = 0.01,
  set_seed = 0, rand_seed = 0)
}
\arguments{
\item{datass}{assessment data}

\item{TAC}{Initial catch level}

\item{Yr_current}{current year}

\item{use_specific_ratios}{switch for specfic ratios}

\item{specific_ratios}{specific ratios}

\item{casal_path}{path to casal.exe}

\item{inputprefix}{CASAL input file prefix}

\item{pop_csl}{CASAL population csl file (name, location???)}

\item{proj_dat}{projection data?}

\item{mvnsamples_dat}{mvn samples}

\item{linux}{run on Windows (default 0) or Linux (1)}

\item{ssb_level_deplet}{depletion level}

\item{ssb_level_target}{target level}

\item{ref_prob_deplet}{reference depletion probability}

\item{ref_prob_target}{reference target probability}

\item{tol}{tolerance}

\item{set_seed}{switch for random number seed}

\item{rand_seed}{random seed}
}
\description{
Update catches, run projections once, return probabilities & extrapolated catch
and check whether probabilities are within tolerance range
}
