% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL3_Run_CASAL.r
\name{run_casal_TAC_finder}
\alias{run_casal_TAC_finder}
\title{CASAL TAC Finder}
\usage{
run_casal_TAC_finder(datass, TAC, Yr_current, use_specific_ratios = 0,
  specific_ratios = 1, casal_path, inputprefix, pop_csl, proj_dat,
  mvnsamples_dat, linux, ssb_level_deplet = 0.2, ssb_level_target = 0.5,
  ref_prob_deplet = 0.1, ref_prob_target = "median", tol = 0.01,
  TAC_multiplier = 1.2, iterations_main = 20, set_seed = 0,
  rand_seed = 0)
}
\arguments{
\item{datass}{assessment data}

\item{TAC}{Initial catch level}

\item{Yr_current}{current year}

\item{use_specific_ratios}{If 1, then use TAC ratios provided by 'specific_ratios'}

\item{specific_ratios}{Specific TAC ratios (if use_specific_ratios=1)}

\item{casal_path}{path to casal.exe}

\item{inputprefix}{CASAL input file prefix}

\item{pop_csl}{CASAL population csl file (name, location???)}

\item{proj_dat}{project data}

\item{mvnsamples_dat}{mvn samples}

\item{linux}{run on Windows (default 0) or Linux (1)}

\item{ssb_level_deplet}{depletion level}

\item{ssb_level_target}{target level}

\item{ref_prob_deplet}{reference depletion probability}

\item{ref_prob_target}{reference target probability}

\item{tol}{tolerance}

\item{TAC_multiplier}{TAC multiplier}

\item{iterations_main}{iterations}

\item{set_seed}{switch for random number seed}

\item{rand_seed}{random seed}
}
\description{
CASAL TAC Finder
}
