% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL3_Run_CASAL.r
\name{run_complete_casal_assessment}
\alias{run_complete_casal_assessment}
\title{Run CASAL assessment (with adjustment for process error, with/witout TAC finder)}
\usage{
run_complete_casal_assessment(ctrl, om, tag, datass, Yr_current, TAC_init = 0,
  intern = TRUE)
}
\arguments{
\item{ctrl}{Parameters for general control}

\item{om}{Parameters for Operating model}

\item{tag}{tag data/parameters}

\item{datass}{Parameters for Assessmment}

\item{Yr_current}{current year}

\item{TAC_init}{Initial catch level}

\item{intern}{CASAL is run internally (default = TRUE) with no output sent
to the R console or it is run with output sent to R console (FALSE, use for
debugging)}
}
\description{
Run CASAL assessment (with adjustment for process error, with/witout TAC finder)
}
\details{
Function to run a complete CASAL assessment loop, including:

- Create input data (source(file="M3_CASAL1_Data.r"))

- Create CASAL input files (create_casal_file_pop, create_casal_file_est,
create_casal_file_out)

- Run CASAL model for point estimatation (run_casal_pointest)

- Draw MVN samples for parameter estimates (run_casal_mvnsamples)

- Find 'optimal' TAC for next year (run_casal_TAC_finder to update catches,
run projections to select optimal TAC that satifies the CCAMLR decision rules)

ctrl Parameters for general control

om	Parameters for Operating model

datass	Parameters for Assessmment

Yr_current	Assessment year, also used to calculate TAC ratios for
all fisheries (if use_specific_ratios=0)

TAC Initial catch level
}
