% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_ages}
\alias{sample_ages}
\title{Age sample}
\usage{
sample_ages(agecomp, sampleN, ageing_cv = 0)
}
\arguments{
\item{agecomp}{Numbers at age in the total catch}

\item{sampleN}{Sample sizes for age sample}

\item{ageing_cv}{Ageing error}
}
\description{
Age sample
}
