% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_catch}
\alias{sample_catch}
\title{Sampling - Catch-at-Age or Length}
\usage{
sample_catch(type, om, sampling, fleet, obs, ff, y, ss, rr)
}
\arguments{
\item{type}{sampling type}

\item{om}{operating model object}

\item{sampling}{sampling object}

\item{fleet}{fleet object}

\item{obs}{observations?}

\item{ff}{fleets???}

\item{y}{year???}

\item{ss}{ss???}

\item{rr}{region???}
}
\description{
Sampling - Catch-at-Age or Length
}
