% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_cpue}
\alias{sample_cpue}
\title{Sampling - Catch rates}
\usage{
sample_cpue(pop, sampling, fleet, ff, y, ss, rr)
}
\arguments{
\item{pop}{population object}

\item{sampling}{sampling object}

\item{fleet}{fleet object}

\item{ff}{fleets???}

\item{y}{year???}

\item{ss}{ss???}

\item{rr}{region???}
}
\description{
Sampling - Catch rates
}
