% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_lengths}
\alias{sample_lengths}
\title{Size sample}
\usage{
sample_lengths(ages1, lenbins, agecomp, growth)
}
\arguments{
\item{ages1}{Number at age of the sample}

\item{lenbins}{Length bins by year, season and region}

\item{agecomp}{Number at age of the sample}

\item{growth}{growth parameters}
}
\description{
Sample lengths
}
