% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_survey}
\alias{sample_survey}
\title{Sampling}
\usage{
sample_survey(type, om, pop, sampling, fleet, obs, ff, y, ss, rr)
}
\arguments{
\item{type}{sampling type}

\item{om}{operating model object}

\item{pop}{population object}

\item{sampling}{sampling object}

\item{fleet}{fleet object}

\item{obs}{observations?}

\item{ff}{fleets???}

\item{y}{year???}

\item{ss}{ss???}

\item{rr}{region???}
}
\description{
Sampling function
}
