% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_OM.r
\name{setup_om_objects}
\alias{setup_om_objects}
\title{Setup operating model objects}
\usage{
setup_om_objects(para)
}
\arguments{
\item{para}{list of model parameters.}
}
\description{
Setup operating model objects ('res' with pop, fleet and obs).
}
