% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{SR}
\alias{SR}
\title{Stock-recruitment relationship}
\usage{
SR(pin, mu, sigma, SSB0 = 1, SSB = 0, rec_series = 0, h = 1)
}
\arguments{
\item{pin}{recruitment input parameters}

\item{mu}{mean recruitment}

\item{sigma}{sd/mu}

\item{SSB0}{SSB0}

\item{SSB}{SSB}

\item{rec_series}{A recruitment series (vector)}

\item{h}{steepness}
}
\description{
Stock-recruitment relationship
}
