% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological_processes.R
\name{stock_recruit}
\alias{stock_recruit}
\title{Stock-recruitment relationship}
\usage{
stock_recruit(method, mu, sigma, SSB0 = 1, SSB = 0, rec_series = 0,
  h = 1, year, rec_min = 1000)
}
\arguments{
\item{method}{recruitment input parameters}

\item{mu}{mean recruitment}

\item{sigma}{sd/mu}

\item{SSB0}{SSB0}

\item{SSB}{SSB}

\item{rec_series}{A recruitment series (vector)}

\item{h}{steepness}

\item{year}{model year (if}

\item{rec_min}{minimum recruitment (default=1000)}
}
\description{
Stock-recruitment relationship
}
