% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging.R
\name{tag_binom}
\alias{tag_binom}
\title{Tag binomial}
\usage{
tag_binom(dat1, prob1, dims)
}
\arguments{
\item{dat1}{Data}

\item{prob1}{Probabilities}

\item{dims}{dimensions}
}
\description{
tag_binom applies to every level of dimension (by units,area,yeartag,areatag).
}
