% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL3_Run_CASAL.r
\name{update_casal_data}
\alias{update_casal_data}
\title{Update CASAL data}
\usage{
update_casal_data(datass, Yr_current)
}
\arguments{
\item{datass}{assessment data}

\item{Yr_current}{current year}
}
\description{
Update CASAL data
}
