% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL1_Data.r
\name{update_casal_file_names}
\alias{update_casal_file_names}
\title{Update CASAL file names}
\usage{
update_casal_file_names(params)
}
\arguments{
\item{params}{Should we call this be para for consistency?}
}
\description{
Update CASAL file names
}
