% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_CASAL3_Run_CASAL.r
\name{update_future_catches}
\alias{update_future_catches}
\title{Update future catches in population.csl}
\usage{
update_future_catches(datass, TAC, Yr_current, use_specific_ratios = 0,
  specific_ratios = 1, casal_path = "", pop_csl = "population.csl")
}
\arguments{
\item{datass}{assessment data}

\item{TAC}{Initial catch level}

\item{Yr_current}{current year}

\item{use_specific_ratios}{If 1, then use TAC ratios provided by 'specific_ratios'}

\item{specific_ratios}{Specific TAC ratios (if use_specific_ratios=1)}

\item{casal_path}{path to casal.exe}

\item{pop_csl}{CASAL population csl file (name, location???)}
}
\description{
Update future catches in population.csl
}
